app/requirements.txt
flask==2.0.1
werkzeug==2.0.1
markdown==3.5.1
python-frontmatter==1.0.0 